Unify Jackson Databind test in a single test.
