+++ vertx-core-jackson-v3/src/test/java/io/vertx/tests/jacksonv3/JacksonDatabindTest.java
@@ -23,6 +23,7 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.core.json.jackson.v3.DatabindCodec;
 import io.vertx.test.core.TestUtils;
 import io.vertx.test.core.VertxTestBase;
+import io.vertx.tests.json.jackson.JacksonDatabindTestBase;
 import org.junit.Test;
 
 import java.io.IOException;
@@ -37,8 +38,17 @@ import static java.time.format.DateTimeFormatter.ISO_INSTANT;
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
  */
-public class JacksonDatabindTest extends VertxTestBase {
+public class JacksonDatabindTest extends JacksonDatabindTestBase {
 
+  @Override
+  protected String writeValueAsString(Object o) throws Exception {
+    return DatabindCodec.mapper().writeValueAsString(o);
+  }
+
+  @Override
+  protected <T> T readValue(String content, Class<T> valueType) throws Exception {
+    return DatabindCodec.mapper().readValue(content, valueType);
+  }
 
   @Test
   public void testGetMapper() {
@@ -71,174 +81,6 @@ public class JacksonDatabindTest extends VertxTestBase {
     assertFalse(incorrect.get(0) instanceof Pojo);
     assertTrue(incorrect.get(0) instanceof Map);
     assertEquals(original.value, ((Map) (incorrect.get(0))).get("value"));
-  }
-
-  @Test
-  public void testInstantDecoding() {
-    Pojo original = new Pojo();
-    original.instant = Instant.from(ISO_INSTANT.parse("2018-06-20T07:25:38.397Z"));
-    Pojo decoded = Json.decodeValue("{\"instant\":\"2018-06-20T07:25:38.397Z\"}", Pojo.class);
-    assertEquals(original.instant, decoded.instant);
-  }
-
-  @Test
-  public void testNullInstantDecoding() {
-    Pojo original = new Pojo();
-    Pojo decoded = Json.decodeValue("{\"instant\":null}", Pojo.class);
-    assertEquals(original.instant, decoded.instant);
-  }
-
-  @Test
-  public void testBytesDecoding() {
-    Pojo original = new Pojo();
-    original.bytes = TestUtils.randomByteArray(12);
-    Pojo decoded = Json.decodeValue("{\"bytes\":\"" + TestUtils.toBase64String(original.bytes) + "\"}", Pojo.class);
-    assertArrayEquals(original.bytes, decoded.bytes);
-  }
-
-  @Test
-  public void testNullBytesDecoding() {
-    Pojo original = new Pojo();
-    Pojo decoded = Json.decodeValue("{\"bytes\":null}", Pojo.class);
-    assertEquals(original.bytes, decoded.bytes);
-  }
-
-
-  @Test
-  public void testJsonArrayDeserializer() throws Exception {
-    String jsonArrayString = "[1, 2, 3]";
-    JsonArray jsonArray = DatabindCodec.mapper().readValue(jsonArrayString, JsonArray.class);
-
-    assertEquals(3, jsonArray.size());
-    assertEquals(new JsonArray().add(1).add(2).add(3), jsonArray);
-
-  }
-
-
-  @Test
-  public void testJsonObjectDeserializer() throws Exception {
-
-    String jsonObjectString = "{\"key1\": \"value1\", \"key2\": \"value2\", \"key3\": \"value3\"}";
-
-    JsonObject jsonObject = DatabindCodec.mapper().readValue(jsonObjectString, JsonObject.class);
-
-    assertEquals("value1", jsonObject.getString("key1"));
-    assertEquals("value2", jsonObject.getString("key2"));
-
-  }
-
-  @Test
-  public void testJsonObjectSerializer() throws Exception {
-
-    ObjectMapper objectMapper = DatabindCodec.mapper();
-
-    JsonObject jsonObject = new JsonObject();
-    jsonObject.put("key1", "value1");
-    jsonObject.put("key2", "value2");
-    jsonObject.put("key3", "value3");
-
-    String jsonString = objectMapper.writeValueAsString(jsonObject);
-
-    assertEquals("{\"key1\":\"value1\",\"key2\":\"value2\",\"key3\":\"value3\"}", jsonString);
-  }
-
-  @Test
-  public void testJsonArraySerializer() throws Exception {
-
-
-    JsonArray jsonArray = new JsonArray();
-    jsonArray.add("value1");
-    jsonArray.add("value2");
-    jsonArray.add("value3");
-
-    String jsonString = DatabindCodec.mapper().writeValueAsString(jsonArray);
-
-    assertEquals("[\"value1\",\"value2\",\"value3\"]", jsonString);
-  }
-
-  @Test
-  public void testInstantSerializer() throws IOException {
-
-    Instant instant = Instant.parse("2023-06-09T12:34:56.789Z");
-
-    String jsonString = DatabindCodec.mapper().writeValueAsString(instant);
-
-    assertEquals("\"2023-06-09T12:34:56.789Z\"", jsonString);
-  }
-
-  @Test
-  public void testInstantDeserializer() throws IOException {
-
-    String jsonString = "\"2023-06-09T12:34:56.789Z\"";
-
-    Instant instant = DatabindCodec.mapper().readValue(jsonString, Instant.class);
-
-    Instant expectedInstant = Instant.parse("2023-06-09T12:34:56.789Z");
-    assertEquals(expectedInstant, instant);
-  }
-
-  @Test
-  public void testByteArraySerializer() throws IOException {
-    byte[] byteArray = "Hello, World!".getBytes();
-
-    String jsonString = DatabindCodec.mapper().writeValueAsString(byteArray);
-
-    String expectedBase64String = Base64.getEncoder().withoutPadding().encodeToString(byteArray);
-    assertEquals("\"" + expectedBase64String + "\"", jsonString);
-  }
-
-  @Test
-  public void testByteArrayDeserializer() throws IOException {
-    String jsonString = "\"SGVsbG8sIFdvcmxkIQ\"";
-
-    byte[] byteArray = DatabindCodec.mapper().readValue(jsonString, byte[].class);
-
-    byte[] expectedByteArray = Base64.getDecoder().decode("SGVsbG8sIFdvcmxkIQ");
-    assertArrayEquals(expectedByteArray, byteArray);
-  }
-
-  @Test
-  public void testBufferSerializer() throws IOException {
-
-    Buffer buffer = Buffer.buffer("Hello, World!");
-
-    String jsonString = DatabindCodec.mapper().writeValueAsString(buffer);
-
-    assertEquals("\"SGVsbG8sIFdvcmxkIQ\"", jsonString);
-  }
-
-  @Test
-  public void testBufferDeserializer() throws IOException {
-    String jsonString = "\"SGVsbG8sIFdvcmxkIQ\"";
-
-    Buffer buffer = DatabindCodec.mapper().readValue(jsonString, Buffer.class);
-
-    Buffer expectedBuffer = Buffer.buffer("Hello, World!");
-    assertEquals(expectedBuffer, buffer);
-  }
-
-  private static class Pojo {
-    @JsonProperty
-    String value;
-    @JsonProperty
-    Instant instant;
-    @JsonProperty
-    byte[] bytes;
-  }
-
-  @Test
-  public void testPrettyPrinting() {
-    JsonObject jsonObject = new JsonObject()
-      .put("key1", "value1")
-      .put("key2", "value2")
-      .put("key3", "value3");
-
-    String compact = Json.encode(jsonObject);
-    String pretty = Json.encodePrettily(jsonObject);
-
-    assertFalse(compact.equals(pretty));
-
-    assertEquals(jsonObject, Json.decodeValue(pretty));
   }
 
   @Test
